Deleted test data

+++ places/models.py
@@ -5,8 +5,8 @@ class Place(models.Model):
     title = models.CharField(max_length=255, unique=True, verbose_name='Название')
     description_short = models.TextField(blank=True, verbose_name='Краткое описание')
     description_long = models.TextField(blank=True, verbose_name='Полное описание')
-    coordinates_lat = models.FloatField(verbose_name='Широта')
-    coordinates_lng = models.FloatField(verbose_name='Долгота')
+    coordinates_lat = models.DecimalField(max_digits=20, decimal_places=15, verbose_name='Широта')
+    coordinates_lng = models.DecimalField(max_digits=20, decimal_places=15, verbose_name='Долгота')
 
     def str(self):
         return self.title
